test: cover Atomizer flow edge cases
